app/requirements.txt
# Core ML and Data Science Libraries
tensorflow==2.15.0
torch==2.1.0
torchvision==0.16.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.3
scipy==1.11.4

# Database Connectivity
mysql-connector-python==8.2.0
PyMySQL==1.1.0
SQLAlchemy==2.0.23

# Web Framework and API
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
requests==2.31.0
PyJWT==2.8.0
bcrypt==4.0.1

# Data Processing and Validation
marshmallow==3.20.1
python-dotenv==1.0.0
Werkzeug==3.0.1

# Utilities
joblib==1.3.2
tqdm==4.66.1
python-dateutil==2.8.2

# Model Optimization
optuna==3.4.0

# Logging and Monitoring
loguru==0.7.2

# File Processing
openpyxl==3.1.2
xlrd==2.0.1 